# remove raw SSE2/NEON implementation from mathfuncs.cpp