@vpisarev 2017-10-11
thank you, @tomoaki0705! I have a similar request about the performance comparison before and after your patch.