@tomoaki0705 2017-10-12
thank you for the response.
Let's finish from the easy part.
Following are the median value from each performance test from ```opencv_perf_core```
I suppose the performance can be called "equal"

Compiler Visual Studio 2013 update 5
OS Windows 7 Ultimate 64bit
CPU Corei7 3720QM 2.60GHz

test case | before | after | x-factor
---- |---- |---- |---- 
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC1)      | 0.091 ms | 0.097 ms |0.94
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC4)      | 0.400 ms | 0.427 ms |0.94
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC1)     | 0.274 ms | 0.295 ms |0.93
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC4)     | 1.419 ms | 1.459 ms |0.97
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC1)    | 0.727 ms | 0.728 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC4)    | 3.430 ms | 3.463 ms |0.99
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC1)    | 3.406 ms | 3.452 ms |0.99
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC4)    |13.872 ms |14.106 ms |0.98
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC1)  | 2.193 ms | 2.183 ms |1.00
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC4)  | 8.744 ms | 8.687 ms |1.01
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC1) | 6.559 ms | 6.514 ms |1.01
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC4) |26.087 ms |26.063 ms |1.00
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC1)|14.705 ms |14.710 ms |1.00
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC4)|58.684 ms |58.666 ms |1.00
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC1)|58.500 ms |58.628 ms |1.00
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC4)|235.133 ms|235.198 ms|1.00

I'll add some tests from ARM devices later