@tomoaki0705 2017-10-14
So, I managed to experiment on some ARM devices.
Here's some data.  Basically, as you concerned, there was a degradation on ARM, so I fixed them.

#### ODROID-X2
test case | before | after | x-factor
---- |---- |---- |---- 
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC1)  | 1.200 ms  |  1.211 ms |0.99
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC4)  | 5.175 ms  |  5.175 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC1) | 3.876 ms  |  3.883 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC4) |15.538 ms  | 15.563 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC1)| 8.758 ms  |  8.752 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC4)|34.963 ms  | 34.980 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC1)|34.972 ms  | 35.009 ms |1.00
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC4)|139.845 ms | 140.626 ms|0.99
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC1)  | 16.777 ms |  16.674 ms  |   1.01
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC4)  | 67.701 ms |  66.952 ms  |   1.01
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC1) | 50.629 ms |  50.198 ms  |   1.01
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC4) | 203.278 ms|  201.073 ms |   1.01
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC1)| 112.662 ms|  111.700 ms |   1.01
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC4)| 458.186 ms|  446.766 ms |   1.03
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC1)| 457.882 ms|  446.844 ms |   1.02
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC4)|1801.110 ms| 1789.127 ms |   1.01

#### Jetson TX1
test case | before | after | x-factor
---- |---- |---- |---- 
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC1)  | 0.274 ms |   0.284 ms  |   0.97
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC4)  | 1.702 ms |   1.692 ms  |   1.01
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC1) | 1.301 ms |   1.366 ms  |   0.95
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC4) | 5.203 ms |   5.200 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC1)| 2.899 ms |   2.902 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC4)|11.749 ms |  11.687 ms  |   1.01
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC1)|11.743 ms |  11.712 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC4)|46.858 ms |  46.741 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC1)  | 8.357 ms |   8.361 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC4)  |33.400 ms |  33.372 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC1) |24.912 ms |  24.938 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC4) |100.220 ms|  99.758 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC1)|56.294 ms |  56.302 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC4)|225.965 ms|  224.816 ms |   1.01
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC1)|224.617 ms|  223.932 ms |   1.00
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC4)|954.012 ms|  924.835 ms |   1.03

#### Jetson TX2
test case | before | after | x-factor
---- |---- |---- |---- 
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC1)  | 0.170 ms |   0.170 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC4)  | 1.058 ms |   1.065 ms  |   0.99
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC1) | 0.793 ms |   0.795 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC4) | 3.301 ms |   3.300 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC1)| 1.845 ms |   1.828 ms  |   1.01
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC4)| 7.407 ms |   7.393 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC1)| 7.379 ms |   7.357 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC4)|29.514 ms |  29.524 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC1)  | 4.365 ms |   3.936 ms  |   1.11
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC4)  |17.588 ms |  15.718 ms  |   1.12
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC1) |13.505 ms |  11.773 ms  |   1.15
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC4) |53.798 ms |  47.211 ms  |   1.14
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC1)|29.304 ms |  26.412 ms  |   1.11
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC4)|117.450 ms|  107.036 ms |   1.10
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC1)|117.981 ms|  106.772 ms |   1.10
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC4)|470.179 ms|  420.751 ms |   1.12

#### ODROID-C2
test case | before | after | x-factor
---- |---- |---- |---- 
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC1)  | 0.883 ms |   0.888 ms  |   0.99
PatchNaNs::OCL_PatchNaNsFixture::(640x480, 32FC4)  | 3.562 ms |   3.573 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC1) | 2.668 ms |   2.685 ms  |   0.99
PatchNaNs::OCL_PatchNaNsFixture::(1280x720, 32FC4) |10.686 ms |  10.730 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC1)| 6.033 ms |   6.044 ms  |   1.00
PatchNaNs::OCL_PatchNaNsFixture::(1920x1080, 32FC4)|24.056 ms |  24.232 ms  |   0.99
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC1)|24.050 ms |  24.183 ms  |   0.99
PatchNaNs::OCL_PatchNaNsFixture::(3840x2160, 32FC4)|97.431 ms |  97.738 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC1)  | 14.322 ms |  14.215 ms  |   1.01
PolarToCart::OCL_PolarToCartFixture::(640x480, 32FC4)  | 57.374 ms |  57.567 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC1) | 42.996 ms |  42.940 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1280x720, 32FC4) | 170.849 ms|  171.608 ms |   1.00
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC1)| 96.106 ms |  95.677 ms  |   1.00
PolarToCart::OCL_PolarToCartFixture::(1920x1080, 32FC4)| 386.188 ms|  387.347 ms |   1.00
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC1)| 386.465 ms|  385.878 ms |   1.00
PolarToCart::OCL_PolarToCartFixture::(3840x2160, 32FC4)|1561.952 ms| 1565.071 ms |   1.00